Build your intitial architecture for your new NOAA Tech Memo or Report

Build your intitial architecture for your new NOAA Tech Memo or Report

buildTM(
  sections = c("frontmatter", "abstract", "introduction", "methods", "results",
    "discussion", "workscited", "workscitedR", "presentation"),
  support_scripts = c("directories", "functions", "dataDL", "data"),
  authors = "",
  title = "",
  styles_reference_pptx = "refppt_nmfs",
  styles_reference_docx = "refdoc_noaa_tech_memo",
  bibliography.bib = "bib_example",
  csl = "apa"
)

Arguments

sections
a string of the different sections of your TM. Default = c("frontmatter", "abstract", "introduction", "methods", "results", "discussion", "workscited", "workscitedR"). Note that frontmatter and workcitedR both have specific templates, and all others are from a blank template.

support_scripts
To make sure we nice and neatly compartemantalize our work, create the below supporting .R files that you will source into your 'run' file. Default = c("functions", "dataDL", "data")

authors
Default = "". Here, add your First Lastname (email).

title
Default = "". Here, put the title of your report.

styles_reference_pptx
A style reference guide from a powerpoint document (.pptx). This pulls the styles from a powerpoint document where you have defined each style. Either use a local document (insert full "path") or some of the pre-made templates ("nmfs"). Default = "nmfs".

styles_reference_docx
A style reference guide from a word document (.docx). This pulls the styles from a word document where you have defined each style. Either use a local document (insert full "path") or some of the pre-made templates ("refdoc_noaa_tech_memo" or "refdoc_fisheries_economics_of_the_us"). Default = "refdoc_noaa_tech_memo".

bibliography.bib
Either use a local document (.bib format; insert full "path") or the example file from the package ("bib_example"). Default = "bib_example".

csl
Citation style. Either use a local document (insert "path") or some of the pre-made templates ("apa", "new-phytologist"). A NOAA TM citation style needs to be created, but until then, the default = "apa".

Examples

# not run:
# buildTM()