Download text from google drive as plain text

Download text from google drive as plain text

googledrive_txt_dl(
  filename_gd = NULL,
  filename_dl = "googledrive_dl_text",
  path = "./",
  verbose = TRUE
)

Arguments

filename_gd
The file path or filename of the google doc from google drive.

filename_dl
The filename you want to download the google doc as. Default = "googledrive_dl_text".

path
The path you want to save this file to. Default = "./"

verbose
Logical, indicating whether to print informative messages (default TRUE).

Value

The plain text from the google doc.

Examples

googledrive::drive_auth() # Using the first token you have.
#> ! Using an auto-discovered, cached token.
#>   To suppress this message, modify your code or options to clearly consent to
#>   the use of a cached token.
#>   See gargle's "Non-interactive auth" vignette for more details:
#>   <https://gargle.r-lib.org/articles/non-interactive-auth.html>
#> i The googledrive package is using a cached token for emilyhmarkowitz@gmail.com.
1
#> [1] 1

txt <- googledrive_txt_dl(filename_gd = "test123123_doc",
                          filename_dl = "test123123_dl",
                          verbose = FALSE)
#> Warning: The `verbose` argument of `some_googledrive_function()` is deprecated as of googledrive 2.0.0.
#> Set `options(googledrive_quiet = TRUE)` to suppress all googledrive messages.
#> For finer control, use `local_drive_quiet()` or `with_drive_quiet()`.
#> googledrive's `verbose` argument will be removed in the future.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.
#> Auto-refreshing stale OAuth token.
txt
#> [1] "This is a test"
# for good file keeping, I'll delete these
file.remove('test123123_dl.txt', 'test123123_dl.zip')
#> [1] TRUE TRUE