Make a string lower case except for stated (and common NOAA) proper nouns.

Make a string lower case except for stated (and common NOAA) proper nouns.

tolower2(str0, capitalizefirst = FALSE, add_cap = "")

Arguments

str0
The text string.

capitalizefirst
Default = FALSE

add_cap
A vector of strings that the user does not want capitalized

Examples

tolower2(str0 = "notice how there are built-in proper nouns are capitalized:
alaska is not in the south atlantic.",
         capitalizefirst = TRUE,
         add_cap = "Proper nouns")

#> [1] "Notice how there are built-in Proper nouns are Alaska is not in the South Atlantic"